Updated count (*) sql to accomodate two different versions of hibernate - it could return either an integer or a long. Also tweaked a method parameters to use the correct Integer/Long class - newer versions of Hibernate throw class cast exceptions when Long is subsituted for integer.

+++ lams_tool_sbmt/src/java/org/lamsfoundation/lams/tool/sbmt/dao/ISubmissionDetailsDAO.java
@@ -58,5 +58,5 @@ public interface ISubmissionDetailsDAO extends IBaseDAO {
 	 */
 	public List getSubmissionDetailsBySession(Long sessionID);
 	
-	public List<SubmissionDetails> getBySessionAndLearner(Long sessionID, Long userID);
+	public List<SubmissionDetails> getBySessionAndLearner(Long sessionID, Integer userID);
 }
